Move all has checks to single has module
