--- src/has/main.ts
@@ -1,3 +1,0 @@
-import has from './has';
-export * from './has';
-export default has;
