+++ tests/has/unit/all.ts
@@ -1,2 +1,1 @@
 import './has';
-import './main';
