--- tests/has/unit/main.ts
@@ -1,16 +1,0 @@
-const { registerSuite } = intern.getInterface('object');
-const { assert } = intern.getPlugin('chai');
-
-import * as main from '../../../src/has/main';
-
-registerSuite('has/main', {
-	'validate api'() {
-		assert.isFunction(main.default);
-		assert.isFunction(main.add);
-		assert.isFunction(main.exists);
-		assert.isFunction(main.normalize);
-		assert.isFunction(main.load);
-		assert.isObject(main.testCache);
-		assert.isObject(main.testFunctions);
-	}
-});
