+++ tests/widget-core/support/sendEvent.ts
@@ -1,5 +1,4 @@
-import has from '../../../src/has/preset';
-import { add } from '../../../src/has/has';
+import has, { add } from '../../../src/has/has';
 import { deepAssign } from '../../../src/core/util';
 import global from '../../../src/shim/global';
 import { assign } from '../../../src/shim/object';
